+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/livedata/ChatDomainImpl.kt
@@ -23,7 +23,6 @@ import io.getstream.chat.android.livedata.controller.QueryChannelsController
 import io.getstream.chat.android.livedata.controller.QueryChannelsControllerImpl
 import io.getstream.chat.android.livedata.controller.ThreadController
 import io.getstream.chat.android.livedata.controller.ThreadControllerImpl
-import io.getstream.chat.android.livedata.usecase.UseCaseHelper
 import io.getstream.chat.android.livedata.utils.Event
 import io.getstream.chat.android.livedata.utils.RetryPolicy
 import io.getstream.chat.android.livedata.utils.toLiveDataRetryPolicy
@@ -71,9 +70,6 @@ internal class ChatDomainImpl internal constructor(internal val chatDomainStateF
 
     override val user: LiveData<User?> = chatDomainStateFlow.user.asLiveData()
 
-    /** a helper object which lists all the initialized use cases for the chat domain */
-    override val useCases: UseCaseHelper = UseCaseHelper(this)
-
     /** if the client connection has been initialized */
     override val initialized: LiveData<Boolean> = chatDomainStateFlow.initialized.asLiveData()
 
@@ -114,6 +110,7 @@ internal class ChatDomainImpl internal constructor(internal val chatDomainStateF
      */
     override val errorEvents: LiveData<Event<ChatError>> = chatDomainStateFlow.errorEvents.map(::Event).asLiveData()
     override val typingUpdates: LiveData<TypingEvent> = chatDomainStateFlow.typingUpdates.asLiveData()
+
     /** The retry policy for retrying failed requests */
     override var retryPolicy: RetryPolicy
         get() = chatDomainStateFlow.retryPolicy.toLiveDataRetryPolicy()