+++ stream-chat-android/src/test/java/com/getstream/sdk/chat/viewmodel/ChannelHeaderViewModelTest.kt
@@ -48,7 +48,7 @@ internal class ChannelHeaderViewModelTest {
     @Test
     fun `Should notify about a new channel`() {
         val channel: Channel = mock()
-        whenever(channelController.channelData) doReturn MutableLiveData(mock())
+        whenever(channelController.offlineChannelData) doReturn MutableLiveData(mock())
         whenever(channelController.toChannel()) doReturn channel
         val channelHeaderViewModel = ChannelHeaderViewModel(CID, chatDomain = chatDomain)
         val mockObserver: Observer<Channel> = spy()